moved functions dedicated to NRB processing from ancillary to nrb module
